requirements.txt
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.13.0
azure-ai-documentintelligence==1.0.2
azure-core==1.39.0
azure-identity==1.25.3
beautifulsoup4==4.14.3
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cobble==0.1.4
colorama==0.4.6
coloredlogs==15.0.1
cryptography==47.0.0
customtkinter==5.2.2
darkdetect==0.8.0
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
flatbuffers==25.12.19
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.13
isodate==0.7.2
jiter==0.14.0
lxml==6.1.0
magika==0.6.3
mammoth==1.11.0
markdownify==1.2.2
markitdown==0.1.5
mpmath==1.3.0
msal==1.36.0
msal-extensions==1.3.1
numpy==2.4.4
olefile==0.47
onnxruntime==1.20.1
openai==2.32.0
openpyxl==3.1.5
packaging==26.2
pandas==3.0.2
pdfminer.six==20251230
pdfplumber==0.11.9
pefile==2024.8.26
pillow==12.2.0
protobuf==7.34.1
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pydub==0.25.1
pyinstaller==6.20.0
pyinstaller-hooks-contrib==2026.4
PyJWT==2.12.1
pypdfium2==5.7.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-pptx==1.0.2
pywin32-ctypes==0.2.3
requests==2.33.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SpeechRecognition==3.16.1
sympy==1.14.0
tkinterdnd2-universal==1.7.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
xlrd==2.0.2
xlsxwriter==3.2.9
youtube-transcript-api==1.0.3
